delete now works yay

+++ src/main/java/uk/ac/ucl/servlets/DeleteItemServlet.java
@@ -12,27 +12,30 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/deleteItem.html")
-public class DeleteItem extends HttpServlet
+@WebServlet("/runDeleteItem.html")
+public class DeleteItemServlet extends HttpServlet
 {
 
-  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException
+  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException
   {
     // Get the data from the model
     Model model = ModelFactory.getModel();
-
+  
     // Then add the data to the request object that will be sent to the Java Server Page, so that
     // the JSP can access the data (a Java data structure).
-    String idRequest = request.getParameter("id");
-    int id = Integer.parseInt(idRequest);
-
-    // delete the selected ID:
-    model.deleteItem(id);
-
+    String idRequest = (String) request.getParameter("deleteid");
+    if (idRequest != null) {
+      int id = Integer.parseInt(idRequest);
+  
+      // delete the selected ID:
+      model.deleteItem(id);
+    }
+  
     // Invoke the JSP.
     // A JSP page is actually converted into a Java class, so behind the scenes everything is Java.
     ServletContext context = getServletContext();
-    RequestDispatcher dispatch = context.getRequestDispatcher("/deleteItem.jsp");
+    RequestDispatcher dispatch = context.getRequestDispatcher("/deleteResult.jsp");
     dispatch.forward(request, response);
   }
+  
 }
